refactor(test): use Typesafe Config system property for DB engine override
Replace withDbEngineOverride() CLI injection with Gradle systemProperty
'storage.db.engine' which ConfigFactory.load() merges automatically.
Rename TestEnv back to TestConstants to minimize review diff.
